Add minimal support for sklearn 1.8.0rc1
scikit-learn 1.8.0rc1 is out, and the final 1.8.0 release will come out
after we do our next cuml release. We don't have capacity or time to
ensure _all_ the upstream sklearn tests pass, but we do want at least
the bare minimum of `cuml.accel` to still work with the upcoming sklearn
1.8.0 release.

This PR defines success in that as:

- `python/cuml/cuml_accel_tests` tests pass
- `python/cuml/tests/test_sklearn_import_export.py` tests pass

We also update our docs to note the upper bound of good testing as
sklearn 1.7 to hedge against other breakages we don't have time to
resolve for a sklearn version that hasn't been released yet.

A summary of fixes

- Skipping `hdbscan` and `umap` tests on sklearn 1.8. The upstream
  packages are not compatible with this version. To make this sane I
  combined the hdbscan integration tests into one file, there's no
  reason to have 2.
- Updating `LogisticRegression` since the `penalty` arg is deprecated
- Dropping a bad `ElasticNet` test for `warm_start` behavior. This
  wasn't actually testing that `warm_start` worked, and improvements to
  the upstream solver made the test invalid.
- Updating our proxied metadata requests methods to the new version used
  by sklearn 1.8.
- Skipping a few other tests on sklearn 1.8 as internal methods have
  changed.
